Changes so maintenance mode response should now be shown during building

+++ app/jobs/MediaItemLiveEmailJob.php
@@ -3,6 +3,7 @@
 use DB;
 use Carbon;
 use EmailHelpers;
+use DebugHelpers;
 use Log;
 use uk\co\la1tv\website\models\MediaItem;
 use uk\co\la1tv\website\models\EmailTasksMediaItem;
@@ -13,40 +14,40 @@ class MediaItemLiveEmailsJob {
 	// expects data to contain "mediaItemId" of the media item that has just gone live
 	public function fire($job, $data) {	
 	
+		if (!DebugHelpers::shouldSiteBeLive()) {
+			$job->release(); // put the job back on the queue
+			return;
+		}
+		
+		// remove the job from the queue to make sure it only runs once even if an exception is thrown
+		$job->delete();
+	
 		$mediaItemId = $data['mediaItemId'];
 		Log::info("Starting job to send email for media item with id ".$mediaItemId." which has gone live.");
 		
-		// only attempt to run the job once
-		if ($job->attempts() <= 1) {
-			
-			// retrieve the media item
-			$mediaItem = DB::transaction(function() use (&$mediaItemId) {
-				$mediaItem = MediaItem::accessible()->whereHas("liveStreamItem", function($q) {
-					$q->accessible()->live();
-				})->whereHas("emailTasksMediaItem", function($q2) {
-					$q2->where("message_type_id", EmailHelpers::getMessageTypeIds()['liveNow'])->where("created_at", ">=", Carbon::now()->subMinutes(15));
-				}, "=", 0)->where("email_notifications_enabled", true)->where("id", $mediaItemId)->lockForUpdate()->first();
-				
-				if (!is_null($mediaItem)) {
-					$emailTask = new EmailTasksMediaItem(array(
-						"message_type_id"	=> EmailHelpers::getMessageTypeIds()['liveNow']
-					));
-					// create an entry in the tasks table for the emails that are going to be sent
-					$mediaItem->emailTasksMediaItem()->save($emailTask);
-				}
-				
-				return $mediaItem;
-			});
+		// retrieve the media item
+		$mediaItem = DB::transaction(function() use (&$mediaItemId) {
+			$mediaItem = MediaItem::accessible()->whereHas("liveStreamItem", function($q) {
+				$q->accessible()->live();
+			})->whereHas("emailTasksMediaItem", function($q2) {
+				$q2->where("message_type_id", EmailHelpers::getMessageTypeIds()['liveNow'])->where("created_at", ">=", Carbon::now()->subMinutes(15));
+			}, "=", 0)->where("email_notifications_enabled", true)->where("id", $mediaItemId)->lockForUpdate()->first();
 			
 			if (!is_null($mediaItem)) {
-				EmailHelpers::sendMediaItemEmail($mediaItem, 'LA1:TV Live Now With "{title}"', "Live now!", "We are streaming live right now!");
+				$emailTask = new EmailTasksMediaItem(array(
+					"message_type_id"	=> EmailHelpers::getMessageTypeIds()['liveNow']
+				));
+				// create an entry in the tasks table for the emails that are going to be sent
+				$mediaItem->emailTasksMediaItem()->save($emailTask);
 			}
 			
-		}
+			return $mediaItem;
+		});
 		
-		// remove the job from the queue
+		if (!is_null($mediaItem)) {
+			EmailHelpers::sendMediaItemEmail($mediaItem, 'LA1:TV Live Now With "{title}"', "Live now!", "We are streaming live right now!");
+		}
 		Log::info("Finished job to send email for media item with id ".$mediaItemId." which has gone live.");
-		$job->delete();
 	}
 	
 }
